requirements.txt
django>=6.0,<7.0
django-cors-headers>=4.3.0,<5.0.0
django-ratelimit>=4.1.0,<5.0.0
pandas>=2.2.0,<3.0.0
numpy>=2.1.0,<3.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
openai>=1.51.0,<2.0.0
python-dateutil>=2.9.0,<3.0.0
email-validator>=2.0.0,<3.0.0
cryptography>=42.0.0,<46.0.0
gunicorn>=22.0.0,<24.0.0